Refactor saltboot classes under saltboot package
