+++ java/code/src/com/redhat/rhn/domain/formula/FormulaFactory.java
@@ -21,8 +21,8 @@ import com.redhat.rhn.common.validator.ValidatorError;
 import com.redhat.rhn.common.validator.ValidatorException;
 import com.redhat.rhn.domain.dto.EndpointInfo;
 import com.redhat.rhn.domain.dto.FormulaData;
-import com.redhat.rhn.domain.image.SaltbootException;
-import com.redhat.rhn.domain.image.SaltbootUtils;
+import com.suse.manager.saltboot.SaltbootException;
+import com.suse.manager.saltboot.SaltbootUtils;
 import com.redhat.rhn.domain.server.MinionServer;
 import com.redhat.rhn.domain.server.MinionServerFactory;
 import com.redhat.rhn.domain.server.Pillar;
