+++ java/code/src/com/suse/manager/saltboot/SaltbootException.java
@@ -1,4 +1,4 @@
-package com.redhat.rhn.domain.image;
+package com.suse.manager.saltboot;
 
 public class SaltbootException extends RuntimeException {
